requirements.txt
accelerate==1.0.1
diffusers==0.30.2
keras==3.6.0
numpy==1.26.4
open_clip_torch==2.20.0
pandas==2.3.3
Pillow==12.0.0
PyYAML==6.0.3
safetensors==0.7.0
salesforce-lavis==1.0.2
scikit-image==0.25.2
tensorflow==2.16.1
tf-keras
torch==2.4.0
tqdm==4.66.5